# Fix LD2450 excessive CPU usage and redundant sensor updates