# What does this implement/fix?

This PR significantly reduces the CPU usage of the LD2450 radar component from 18.5% to 4.0% by implementing several performance optimizations. The component was previously consuming excessive CPU time and generating redundant network traffic by publishing identical sensor values repeatedly.

**Background**: The LD2450 radar sensor (used in devices like Apollo R Pro 1) was flooding networks with redundant updates, publishing the same sensor values multiple times per second even when nothing changed. This caused unnecessary network congestion and excessive CPU usage on the ESP device.

## Key Optimizations

1. **Change Detection**: Implements caching to only publish sensor values when they actually change, eliminating redundant updates
2. **Mathematical Optimizations**: Replaces expensive `pow()` function calls with simple multiplication for distance calculations
3. **String Operation Elimination**: Replaces string-based hex conversion with efficient bitwise operations
4. **Early Throttle Check**: Moves throttling logic before buffer validation to avoid unnecessary processing

## Performance Results

**Before optimizations:**
- CPU usage: 18.5% (11,073ms out of 60,000ms)
- Average processing time: 1.95ms per update
- Update frequency: ~94 updates/second

**After optimizations:**
- CPU usage: 4.0% (2,400ms out of 60,000ms)  
- Average processing time: 0.47ms per update
- Update frequency: ~86 updates/second

**Overall improvement: 78% reduction in CPU usage**

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

- fixes #(issue number if applicable)

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- N/A (no documentation changes required)

## Test Environment

- [x] ESP32
- [ ] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# No configuration changes required - optimizations are internal
ld2450:
  - id: ld2450_radar
    uart_id: ld2450_uart
    throttle: 1000ms

sensor:
  - platform: ld2450
    ld2450_id: ld2450_radar
    target_count:
      name: Presence Target Count
    # ... other sensors
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).

## Technical Details

### Change Detection Implementation
- Added cache structures to track previous sensor values
- Only publishes when values differ from cached values
- Uses proper sentinel values that don't conflict with valid sensor ranges

### Mathematical Optimizations
- Replaced `pow(x, 2)` with `x * x` for distance calculations
- Reuses already decoded coordinate values instead of decoding twice

### String Operation Elimination
- Replaced `str_snprintf()` and `std::stoi()` based hex conversion with direct bitwise operations
- Eliminated temporary string allocations in the data processing path

### Code Safety
- All sentinel values are outside the valid sensor data ranges
- No functional changes to sensor behavior
- Maintains backward compatibility