@kbx81 2025-07-05
I'll test on real hardware this evening